Split release control scope/workflow into two tabs/steps (#831)
