+++ src/apps/ReleaseControl/components/Form/ReleaseControlProcessForm.tsx
@@ -1,8 +1,7 @@
-import { Button, Progress, SingleSelect } from '@equinor/eds-core-react';
+import { Button, Icon, Progress, SingleSelect } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { useState } from 'react';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import styled from 'styled-components';
 import { PhaseSelect } from '../../../DisciplineReleaseControl/Components/Form/Inputs/PhaseSelect';
 import { getReleaseControlById } from '../../api/releaseControl/Request';
 import { DRCFormAtomApi } from '../../Atoms/formAtomApi';
@@ -13,13 +12,23 @@ import { CreateReleaseControlStepModel } from '../../types/releaseControl';
 import { disciplineReleaseControlFactoryContext } from '../Factory/FactoryComponent';
 import { ReleaseControlSidesheet } from '../sidesheet/ReleaseControlSidesheet';
 import { DescriptionInput, PlannedDueDateInput, ReferencesInput, TitleInput } from './Inputs';
-import { FlexColumn, FormWrapper, Wrapper } from './releaseControlProcessForm.styles';
+import {
+    ActionBar,
+    ButtonContainer,
+    FlexColumn,
+    FormWrapper,
+    NavigationButton,
+    NewStepButton,
+    SelectionRow,
+    Wrapper,
+} from './releaseControlProcessForm.styles';
 import { WorkflowCustomEditor } from './WorkflowEditor/WorkflowCustomEditor';
-import { addStep } from './WorkflowEditor/WorkflowEditorHelpers';
+import { addStep, updateStep } from './WorkflowEditor/WorkflowEditorHelpers';
 
 export const ReleaseControlProcessForm = (): JSX.Element => {
     const { useAtomState, updateAtom } = DRCFormAtomApi;
     const steps = useAtomState(({ workflowSteps }) => workflowSteps ?? []);
+    const step = useAtomState(({ step }) => step ?? 'scope');
 
     const { workflowsQuery, workflowTemplateQuery } = releaseControlQueries;
 
@@ -45,38 +54,42 @@ export const ReleaseControlProcessForm = (): JSX.Element => {
     return (
         <Wrapper>
             <FormWrapper>
-                <FlexColumn>
-                    General info
-                    <TitleInput />
-                    <DescriptionInput />
-                    <PlannedDueDateInput />
-                    <PhaseSelect />
-                    <ReferencesInput />
-                </FlexColumn>
-                <FlexColumn>
-                    Workflow
-                    <div style={{ fontWeight: 400, fontSize: '16px' }}>
-                        Select a workflow to start with or create a complete custom flow.
-                    </div>
-                    <SelectionRow>
-                        <SingleSelect
-                            items={workflows?.map((x) => x.name) ?? []}
-                            label="Workflow"
-                            placeholder="Select new or predefined workflow"
-                            size={35}
-                            handleSelectedItemChange={(change) => {
-                                const id = workflows?.find(
-                                    (x) => x.name === change.selectedItem
-                                )?.id;
-                                setValue(id ?? null);
-                            }}
-                        />
-                    </SelectionRow>
-                    <WorkflowCustomEditor />
-                    {steps.length !== 0 && (
-                        <NewStepButton onClick={() => addStep(steps)}>Add step</NewStepButton>
-                    )}
-                </FlexColumn>
+                {step === 'scope' && (
+                    <FlexColumn>
+                        General info
+                        <TitleInput />
+                        <DescriptionInput />
+                        <PlannedDueDateInput />
+                        <PhaseSelect />
+                        <ReferencesInput />
+                    </FlexColumn>
+                )}
+                {step === 'workflow' && (
+                    <FlexColumn>
+                        Workflow
+                        <div style={{ fontWeight: 400, fontSize: '16px' }}>
+                            Select a workflow to start with or create a complete custom flow.
+                        </div>
+                        <SelectionRow>
+                            <SingleSelect
+                                items={workflows?.map((x) => x.name) ?? []}
+                                label="Workflow"
+                                placeholder="Select new or predefined workflow"
+                                size={35}
+                                handleSelectedItemChange={(change) => {
+                                    const id = workflows?.find(
+                                        (x) => x.name === change.selectedItem
+                                    )?.id;
+                                    setValue(id ?? null);
+                                }}
+                            />
+                        </SelectionRow>
+                        <WorkflowCustomEditor />
+                        {steps.length !== 0 && (
+                            <NewStepButton onClick={() => addStep(steps)}>Add step</NewStepButton>
+                        )}
+                    </FlexColumn>
+                )}
             </FormWrapper>
             <SubmitButtonBar />
         </Wrapper>
@@ -84,9 +97,11 @@ export const ReleaseControlProcessForm = (): JSX.Element => {
 };
 
 export const SubmitButtonBar = (): JSX.Element => {
-    const { useIsValid } = DRCFormAtomApi;
+    const { useIsValid, useAtomState } = DRCFormAtomApi;
 
     const isValid = useIsValid();
+
+    const step = useAtomState(({ step }) => step ?? 'scope');
 
     const swapComponent = disciplineReleaseControlFactoryContext.useAtomState(
         ({ swapComponent }) => swapComponent
@@ -124,6 +139,17 @@ export const SubmitButtonBar = (): JSX.Element => {
 
     return (
         <ActionBar>
+            <NavigationButton>
+                {step === 'workflow' && (
+                    <Button variant="outlined" onClick={() => updateStep('scope')}>
+                        <Icon
+                            name={'chevron_left'}
+                            color={tokens.colors.interactive.primary__resting.rgba}
+                        />
+                        Back to select scope
+                    </Button>
+                )}
+            </NavigationButton>
             <ButtonContainer>
                 {isLoading ? (
                     <Button variant="ghost_icon" disabled>
@@ -131,6 +157,15 @@ export const SubmitButtonBar = (): JSX.Element => {
                     </Button>
                 ) : (
                     <>
+                        {step === 'scope' && (
+                            <Button onClick={() => updateStep('workflow')}>
+                                Next: select workflow
+                                <Icon
+                                    name={'chevron_right'}
+                                    color={tokens.colors.text.static_icons__primary_white.hex}
+                                />
+                            </Button>
+                        )}
                         <Button disabled={!isValid} onClick={() => onMutate(false)}>
                             Submit
                         </Button>
@@ -148,31 +183,3 @@ export const SubmitButtonBar = (): JSX.Element => {
         </ActionBar>
     );
 };
-
-export const ButtonContainer = styled.div`
-    flex-direction: row;
-    gap: 0.5em;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    padding: 1em;
-`;
-
-export const ActionBar = styled.div`
-    height: 64px;
-    width: 100%;
-    border: 1px solid ${tokens.colors.interactive.disabled__border.hex};
-    background-color: white;
-`;
-
-export const SelectionRow = styled.div`
-    display: flex;
-    flex-direction: row;
-`;
-
-export const NewStepButton = styled(Button)`
-    margin-bottom: 20px;
-    margin-left: 60px;
-    margin-top: 16px;
-    width: 100px;
-`;
